requirements.txt
numpy>=1.21.0
opencv-python>=4.5.4
Pillow>=9.0.0
scikit-learn>=1.0.2
scipy>=1.7.3
pandas>=1.3.5
matplotlib>=3.5.1
tqdm>=4.62.3
pyyaml>=6.0
joblib>=1.1.0
